# docs/bounded-contexts/catalog/api.yaml
openapi: 3.0.3
info:
  title: 商品限定上下文 API
  description: 类别、商品（SPU）、规格维度与选项、SKU 的 REST 契约。Step Definitions 与实现均按此规范。
  version: 0.1.0
paths:
...
  /api/products/{id}:
    get:
      summary: 按 ID 查商品详情
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 商品详情
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '404':
          description: 商品不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    put:
      summary: 修改商品
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductUpdate'
      responses:
        '200':
          description: 修改成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '404':
          description: 商品不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      summary: 删除商品
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: 删除成功，无响应体
        '404':
          description: 商品不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # ---------- 规格维度与选项 ----------
...
  /api/products/{spuId}/images:
    post:
      summary: 为产品添加展示图（不指定选项）
      description: 添加产品级展示图，不关联任何规格选项。
      parameters:
        - name: spuId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductImageCreate'
      responses:
        '201':
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductImage'
        '404':
          description: SPU 不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '400':
          description: 业务校验失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    get:
      summary: 查询某 SPU 的产品级展示图列表
      description: 返回不关联任何选项的展示图，按 sortOrder 排序。
      parameters:
        - name: spuId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 展示图列表（按 sortOrder 排序）
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProductImage'
        '404':
          description: SPU 不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/products/{spuId}/images/{imageId}:
    delete:
      summary: 删除产品级展示图
      parameters:
        - name: spuId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: imageId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: 删除成功
        '404':
          description: 展示图不存在或不属于该 SPU
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # ---------- 选项展示图 ----------
...
  /api/products/{spuId}/dimensions/{dimensionId}/options/{optionId}/images/{imageId}:
    delete:
      summary: 删除展示图
      parameters:
        - name: spuId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: dimensionId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: optionId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: imageId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: 删除成功
        '404':
          description: 展示图不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # ---------- SKU ----------
  /api/skus/{id}:
    get:
      summary: 按 ID 查 SKU 详情（供 Order 等下游微服务调用）
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: SKU 详情
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Sku'
        '404':
          description: SKU 不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/products/{spuId}/skus:
    post:
      summary: 在 SPU 下创建 SKU（选齐必填维度 Option + 价格）
      parameters:
        - name: spuId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SkuCreate'
      responses:
        '201':
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Sku'
        '404':
          description: SPU 不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '400':
          description: 业务校验失败（未选齐必填维度、Option 不属于该 SPU、价格不能为负等）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    get:
      summary: 按 SPU 查该 SPU 下所有 SKU
      parameters:
        - name: spuId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: SKU 列表
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Sku'
        '404':
          description: SPU 不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/products/{spuId}/skus/{skuId}:
    get:
      summary: 按 SPU 与 SKU ID 查 SKU 详情
      parameters:
        - name: spuId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: skuId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: SKU 详情
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Sku'
        '404':
          description: SKU 不存在或不属于该 SPU
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    put:
      summary: 修改 SKU（如价格）
      parameters:
        - name: spuId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: skuId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SkuUpdate'
      responses:
        '200':
          description: 修改成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Sku'
        '404':
          description: SKU 不存在或不属于该 SPU
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '400':
          description: 业务校验失败（如价格不能为负）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      summary: 删除 SKU
      parameters:
        - name: spuId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: skuId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: 删除成功
        '404':
          description: SKU 不存在或不属于该 SPU
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
components:
  schemas:
...
    Product:
      type: object
      required: [id, categoryId, name]
      properties:
        id:
          type: integer
          format: int64
        categoryId:
          type: integer
          format: int64
        name:
          type: string
        description:
          type: string
          nullable: true
...
    ProductUpdate:
      type: object
      required: [name]
      properties:
        name:
          type: string
        description:
          type: string
          nullable: true
    Error:
      type: object
      properties:
        message:
          type: string
          description: 错误说明，便于 Step 断言

    # ---------- 规格维度与选项 ----------
...
    ProductImage:
      type: object
      required: [id, spuId, imageUrl]
      properties:
        id:
          type: integer
          format: int64
        spuId:
          type: integer
          format: int64
        specOptionId:
          type: integer
          format: int64
          nullable: true
          description: 为空表示产品级展示图，不为空表示该选项的展示图
        imageUrl:
          type: string
        sortOrder:
          type: integer
          nullable: true
    ProductImageCreate:
      type: object
      required: [imageUrl]
      properties:
        imageUrl:
          type: string
        sortOrder:
          type: integer
          nullable: true

    # ---------- SKU ----------
    Sku:
      type: object
      required: [id, spuId, priceCents]
      properties:
        id:
          type: integer
          format: int64
        spuId:
          type: integer
          format: int64
        priceCents:
          type: integer
          format: int64
          minimum: 0
        displayName:
          type: string
          nullable: true
        specValues:
          type: array
          description: 所选 Option 信息，便于展示（如 [{ dimensionName, optionValue }]）
          items:
            type: object
            properties:
              dimensionName:
                type: string
              optionValue:
                type: string
    SkuUpdate:
      type: object
      properties:
        priceCents:
          type: integer
          format: int64
          minimum: 0
          description: 新价格（分）
        displayName:
          type: string
          nullable: true
    SkuCreate:
      type: object
      required: [specOptionIds, priceCents]
      properties:
        specOptionIds:
          type: array
          description: 每个必填维度选一 Option 的 ID，顺序不限
          items:
            type: integer
            format: int64
        priceCents:
          type: integer
          format: int64
          minimum: 0
        displayName:
          type: string
          nullable: true